server/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
python-dotenv==0.19.1
openai >=1.6.1
requests >= 2.31.0
gunicorn==20.1.0
werkzeug == 2.3.7
nltk==3.8.1